pyproject.toml
[project]
name = "wormhole"
version = "0.1.0"
description = "Agent2Agent Protocol - An open protocol enabling communication between AI agents"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "wormhole", email = "hello@wormholeportal.io" }
]
dependencies = [
    "httpx>=0.27.2",
    "httpx-sse>=0.4.0",
    "jwcrypto>=1.5.6",  
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.1",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.34.0",
    "click>=8.1.8",
    "python-dotenv>=1.1.0",
    "openai",
    "openai-agents",
    "google-adk==0.4.0",
    "asyncclick",
    "litellm",
    "google-genai"
]

[project.optional-dependencies]
dev = ["pytest>=8.3.5", "pytest-mock>=3.14.0", "ruff>=0.11.2"]
adk = ["google-adk==0.4.0", "litellm"]
oai = ["openai","openai-agents"]

[project.urls]
Homepage = "https://github.com/wormholeportal"
Documentation = "https://github.com/wormholeportal/Wormhole"

[tool.hatch.build.targets.wheel]
packages = ["wormhole"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"